♻️ zv: variants for `ObjectPath`, `Signature`
`ObjectPath` and `Signature` are a "special" kind of string in
DBus. On the wire it gets encoded with a special type (`o` or
`s`) and is guaranteed to be formated as valid DBus object path
or DBus type signature. In order to handle this correctly the
`ObjectPath` and `Signature` implementations need to hint during
serialization that this string is _more_ than _just_ a string.

During deserialization a similiar hint is necessary. `ObjectPath`
and `Signature` need to hint down into the deserializer that
they are expecting something "more special" than a normal string.

So, this change does that hinting by boxing the the `ObjectPath`
and `Signature` during serialization into a newtype struct. This
form of struct in rust is basically purpose built for this use
case and it maps really well to our needs.