🎨 zv: add nom-based signature parser
This change adds a nom-based signature parser. This parser takes
signatures and breaks them out into an Enum that can be easily
dealt with in the serializer and deserializer with straightforward
pattern matching. Because it is a recursive-descent parser it
trivially handles breaking apart nested signatures in a very nice
way.